+++ src/model/VisualState.ts
